Major audio rework
Audio can now be interracted with outside of its server, base on ids or where the ctx.author is connected.
Not all edge-cases have been tested, but no bugs are currently known either.
Do not try to resume audio where it left off when switching vcs, it's a fools errand.
Fixed a formatting bug in Integration, which was dangerously lacking in permission-checking.
Improved error output